[feature] Mint With Signature This commit modifies the mint behaviour to allow minting with a signature. In previous commits, anyone could mint on behalf of a creator. As a result, anyone could attribute unwanted media to a creator. To fix this, the `mint` function has been modified to only allow msg.sender as the creator. To allow 3rd parties to mint on a creator's behalf, the 3rd party must now collect a signature from the creator, approving the media about to be minted. In order to avoid the callstack depth being exceeded, two new structs have been introduced, modifying the Permit method and the mint methods.
EIP712Signature is a struct containing the v, r, s, and deadline values for a signature. MediaData is a struct containing the URIs and hashes for the media. Note that permit now expects a signature passed in as a struct, and mint and mintWithSig both expect a MediaData and EIP712Signature struct as parameters.